+++ Z1R_Tracker/Z1R_Tracker/DungeonData.fs
@@ -30,7 +30,7 @@ module Factoids =
         "When playing 'swordless', the only way to defeat a gleeok is with a wand, and the only way to defeat wizzrobes is with explosions (from bombs or boomstick)."
         "If a dungeon is on screen A15, recorder-ing to that dungeon will cause the whirlwind drop you at B15, as there's not space on A15 to land."
         "If you clear a dungeon room containing traps or bubbles, it will stay clear forever until you leave the dungeon (as traps & bubbles are considered enemies)."
-        "Up-A-continue during a fairy fountain refill returns you to the starting screen with full health."
+        "Up-A-continue/save during a fairy fountain refill returns you to your starting screen on the overworld with full health."
         "Bomb upgrade man never has a bombable north wall (nor do rooms containing dungeon NPCs who give hints)."
         "Standing just halfway out the east door of a dungeon room makes the player immune to some attacks; this 'safety door' strategy can be effective to shoot beam swords at gleeoks or lanmola, for example."
         |]