+++ doc/whats-new.md
@@ -75,7 +75,8 @@ a certain item, but you forget what you set it to.
  - made dungeon LEVEL/BOARD header use Zelda font
  - speaking either "don't care" or "nothing" will mark off an overworld tile using [Speech Recognition](use.md#speech-recognition)
  - when you get a reminder to consider the magical sword, tracker now briefly highlights its location on the map
- - when you get the 8th triforce piece, tracker now briefly highlights the location of dungeon 9
+ - when you get the 8th triforce piece, reminder that dungeon 9 is unlocked now briefly highlights the location of dungeon 9
+ - when you reset the timer, you immediately get put into the 'place the start spot location icon' popup to mark your start screen
 
 ---